style: InputDailyContent 컴포넌트 분리 & zustand 사용
