+++ next.config.mjs
@@ -8,6 +8,9 @@ const nextConfig = {
 
     return config;
   },
+  images: {
+    domains: ['avatars.githubusercontent.com', 'cdn.smarttoday.co.kr'],
+  },
 };
 
 export default nextConfig;
