+++ src/app/(afterLogin)/posts/layout.tsx
