+++ src/app/_components/common/header/ControlHeader.tsx
@@ -1,8 +1,11 @@
-interface Props {
-  isValid: boolean;
-}
+'use client';
 
-export default function ControlHeader({ isValid }: Props) {
+import { useInputStore } from '@/app/_store/inputStore';
+
+export default function ControlHeader() {
+  const { inputText } = useInputStore();
+  const isValid = inputText.length > 0;
+
   return (
     <header className="header justify-between px-[1.6rem]">
       <button type="button" className="font-body">
