+++ src/app/_store/inputStore.ts
@@ -1,0 +1,11 @@
+import { create } from 'zustand';
+
+interface InputState {
+  inputText: string;
+  setInputText: (text: string) => void;
+}
+
+export const useInputStore = create<InputState>((set) => ({
+  inputText: '',
+  setInputText: (text) => set({ inputText: text }),
+}));
